requirements.txt
alembic==1.13.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
flasgger==0.9.7.1
Flask==3.0.2
Flask-Migrate==4.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==7.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.17.0
